requirements.txt
fastapi
uvicorn

jinja2

aiofiles

sqlalchemy
psycopg2-binary

python-dotenv

pydantic[email]

passlib[bcrypt]

python-jose

python-multipart

databases
asyncpg
